[Refactor] 지하철 검색 컴포넌트 의존성 역전 적용 및 데이터 핸들링 로직 분리
